requirements.txt
anyio==3.6.2
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.15
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
